Split up lrmon a bit, share series between plots
